requirements.txt
datasets==2.18.0
fastapi==0.115.4
jieba==0.42.1
nltk==3.8.1
numpy==2.1.1
ollama==0.3.3
peft==0.10.0
py2neo==2021.2.4
pyahocorasick==2.1.0
rouge_chinese==1.0.3
ruamel.base==1.0.0
scikit_learn==1.5.2
seqeval==1.2.2
streamlit==1.32.2
torch==2.4.1+cu124
tqdm==4.66.2
transformers==4.39.0
typer==0.12.3
uvicorn==0.34.0
zhipuai==2.0.1
